Extract desktop IPC contract package
