+++ apps/app/package.json
@@ -18,6 +18,7 @@
     "@bb/agent-providers": "workspace:*",
     "@bb/config": "workspace:*",
     "@bb/core-ui": "workspace:*",
+    "@bb/desktop-contract": "workspace:*",
     "@bb/domain": "workspace:*",
     "@bb/fuzzy-match": "workspace:*",
     "@bb/host-daemon-contract": "workspace:*",
