+++ apps/desktop/src/desktop-browser-view.ts
@@ -12,7 +12,7 @@ import {
   type BbDesktopBrowserState,
   type BbDesktopBrowserViewportBounds,
   type BbDesktopBrowserViewBounds,
-} from "@bb/server-contract";
+} from "@bb/desktop-contract";
 import {
   BB_DESKTOP_BROWSER_OPEN_TAB_CHANNEL,
   BB_DESKTOP_BROWSER_SNAPSHOT_CHANNEL,
@@ -296,7 +296,9 @@ function historyNavigationTargetUrl(
   offset: number,
 ): string | null {
   const history = entry.view.webContents.navigationHistory;
-  const targetEntry = history.getEntryAtIndex(history.getActiveIndex() + offset);
+  const targetEntry = history.getEntryAtIndex(
+    history.getActiveIndex() + offset,
+  );
   return typeof targetEntry?.url === "string" ? targetEntry.url : null;
 }
