+++ apps/desktop/src/preload.ts
@@ -20,7 +20,7 @@ import {
   type BbDesktopPopoutThreadChangedHandler,
   type BbDesktopPopoutUnsubscribe,
   type BbDesktopTheme,
-} from "@bb/server-contract";
+} from "@bb/desktop-contract";
 import {
   BB_DESKTOP_CHECK_FOR_UPDATES_CHANNEL,
   BB_DESKTOP_GET_INFO_CHANNEL,
