+++ packages/desktop-contract/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./browser.js";
+export * from "./info.js";
+export * from "./popout.js";
+export * from "./version-feed.js";
