+++ packages/desktop-contract/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": [
+    "@bb/tsconfig/base.json",
+    "@bb/tsconfig/typecheck-overrides.json"
+  ],
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": ["src", "test", "tests", "fixtures", "stories"]
+}
